Data initialization of Room not working as expected..!
Some data classes updated
Binding real estate repository
TypeConverter introduced
Kotlin delegation related gradle added

+++ app/src/main/java/com/openclassrooms/realestatemanager/data/model/AgentEntity.kt
@@ -3,12 +3,12 @@ package com.openclassrooms.realestatemanager.data.model
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 
-@Entity(
-    tableName = "agents"
-)
+@Entity(tableName = "agents")
 data class AgentEntity(
+
     @PrimaryKey(autoGenerate = true)
-    val id: Long = 0,
+    val id: Int = 0,
+
     val name: String,
     val email: String,
     val photoUrl: String
